Route details (last tour and history list item); Added Fixture; Added name and finished column to route

+++ app/src/main/java/pg/groupproject/aruma/fragments/history/HistoryDetailsAltitudeChartAdapter.java
@@ -5,32 +5,32 @@ import com.robinhood.spark.SparkAdapter;
 import java.util.Random;
 
 public class HistoryDetailsAltitudeChartAdapter extends SparkAdapter {
-    private float[] yData;
+	private double[] yData;
 
-    public HistoryDetailsAltitudeChartAdapter(float[] yData) {
-        this.yData = yData;
-    }
+	public HistoryDetailsAltitudeChartAdapter(double[] yData) {
+		this.yData = yData;
+	}
 
-    public HistoryDetailsAltitudeChartAdapter() {
-        Random random = new Random();
-        yData = new float[50];
-        for (int i = 0, count = yData.length; i < count; i++) {
-            yData[i] = random.nextFloat();
-        }
-    }
+	public HistoryDetailsAltitudeChartAdapter() {
+		Random random = new Random();
+		yData = new double[50];
+		for (int i = 0, count = yData.length; i < count; i++) {
+			yData[i] = random.nextFloat();
+		}
+	}
 
-    @Override
-    public int getCount() {
-        return yData.length;
-    }
+	@Override
+	public int getCount() {
+		return yData.length;
+	}
 
-    @Override
-    public Object getItem(int index) {
-        return yData[index];
-    }
+	@Override
+	public Object getItem(int index) {
+		return yData[index];
+	}
 
-    @Override
-    public float getY(int index) {
-        return yData[index];
-    }
+	@Override
+	public float getY(int index) {
+		return (float) yData[index];
+	}
 }
